requirements.txt
Flask==2.2.3
opencv-python==4.7.0.72
picamera2==1.0.0  # This is for Raspberry Pi Camera
pyserial==3.5
